debug valid_ratio

+++ model.py
@@ -236,14 +236,14 @@ if __name__ == '__main__':
                 model.eval()
                 valid_loss = 0
                 for batch_id, data_batch in enumerate(valid_dataloader):
-                    if batch_id == len(valid_dataloader) * valid_used_ratio:
+                    if batch_id == int(len(valid_dataloader) * valid_used_ratio):
                         break
                     data_batch = data_batch.to(device)
                     y_pred = model(data_batch)
                     loss = criterion(y_pred, labels)
                     valid_loss += loss.item()
 
-                valid_loss = valid_loss/(len(valid_dataloader) * valid_used_ratio)
+                valid_loss = valid_loss/int(len(valid_dataloader) * valid_used_ratio)
                 if valid_loss < lowest_loss:
                     lowest_loss = valid_loss
                     save_checkpoint(epoch, model, optimizer, valid_loss)
